Completed the 2.5 Creating Scripts Step
Completed the 2.5 Creating Scripts Step defined as "A sample script should be created in the project folder. Preferably, it should have its own Scripts folder. The script should derive from MonoBehaviour. The script then should be added to one of the moveable objects." You could control the sphere object  stading alone on the right of the wall using your keyboard.