Completed the 2.6 Applying Forces to RigidBodies
Completed the 2.6 Applying Forces to RigidBodies step defined as "The script should take the RigidBody component of the object and add a certain amount of force to it every frame. The script can acquire the RigidBody via Unity’s GUI by dragging and dropping, outside of the script itself." The cube goes upwards and hits the brick wall placed as a free falling ceiling.