Updated 2.6, Completed 2.7 Adding Textures Step
Updated 2.6 part. Completed the 2.7 Adding Textures Step defined as "Materials should be created using images, which can then be assigned to the Mesh Renderer of any 3D object. The image can be added to any of the fields of the material (Albedo, Metallic, etc.) proving different results". Used both albedo and metallic options to see the effects of textures on the objects.