Completed the 2.8 Prefabs Step
Completed the 2.8 Prefabs Step defined as "Prefabs are an easy way to save a certain object that you created so that it can be used and spawned as necessary. A prefab from any of the movable objects should be created and must have a script that can be used to spawn new objects by pressing a button on the keyboard." You could see and test the prefab added to the Darwin sphere by pressing space button while playing and spawning multiple of the same sphere.