Completed the 2.9 Adding Lighting Step
Completed the 2.9 Adding Lighting Step defined as "Various sources of lighting should be added to the scene. Be mindful that lighting systems work off of the GPU." You could see the point light added to the initial point of Sauron and the directional light added to the initial point of Gandalf.