Completed the 2.10 Modifying the Camera Step
Completed the 2.10 Modifying the Camera Step defined as "The camera should be modified to show a better view of the scene. its position, rotation, scale, Field of View, and more, can all be adjusted inside the scene." You could see a perfect view of the scene where all of the objects are located from the main camera perspective.