Completed the 2.11 Making Objects Respond Step
Completed the 2.11 Making objects respond to inputs step defined as "Games require a player to function. By adding a way to check for inputs inside a script, we can control an object. For this exercise, the object should respond to inputs by adding force in the direction the inputs were received from, preferably by using WASD or arrow keys." You can control Kratos sphere with both your WASD and arrow keys. I also added a Zeus figure as a sandbag for Kratos.