Updated 2.11, Completed 2.12 Camera Following
Updated 2.11. Completed 2.12 Making the Camera Following the Player defined as "Games need to have the camera focus on the play area, so that the player does not miss out on any action. The position of the camera depends on the type of game (First Person Shooters vs. 2D Platformers). The camera in this case should follow the ”player object” that was defined earlier. Even though this can be done via a script that is applied to the camera, there is a much simpler way to accomplish this." You can watch the Kratos sphere as you move it with your WASD or arrow keys from the Display 2 option of the game view. This commit concludes the HW#1.