Created the initial project folder for HW#2
Created the initial project folder and connected it to the GitHub so that the second homework can be easily managed with a clearly working pipeline. Added the water surrounding the islands and the bridges between the islands. These assets were found from the asset store of the Unity in line with the expectations described in the project description.