Created the Term Project: Spacecape
Created the initial base for the term project named Spacecape in order to build upon. Some portion of this base were taken from the HW#2 design. The Spacecape will include a puzzle solving survival theme for the player to run away from zombies to get back to the space station.