Created the Teleport
Created the Teleport from the Space station to the earth so that Player can vander around the no longer unhabitable earth escaping the zombies and trying to collect different objects to get back to the space station. The skybox and the background music of the game will be changed once the player gets to the earth. Both of the background musics are created using artificial intelligence tools to make them comply with the survival theme of the game.