Created the AI Follower Zombie
Created the Follower Zombie using Unity NavMesh bake option and artificial intelligence library to make the zombies follow the player based on their distance with a constant speed. In case the player collides with a zombie on Earth, the zombie will eat the player alive. Thus, the player will lose the game. Also, created the screen to denote the death screen for the case specific to zombies.