Created the Coins
Created the Coins and arranged the interactions between player and these coins. When the player collects 4 coins, the key object will be visible. After collecting a total of 6 coins, the gumball object will be visible to teleport the player into another maze. Included different zombies on different islands and created the interactions between key object and the door to Watterson island.