Completed the General Structure
Completed the General Structure of the term project before finalizing the work. Created the portal teleporting the player from the earth to the space station. Added the sound effects for zombies, coins, and doors. Created the first puzzle of the space station with a sound effect from a God of war game to offer player a choice between two game objects. Added coin number to the canvas so that player can keep track of the coin value and the required number of coins to complete each step. Also, modified the maze structure on Earth.