Finalized the Term Project
Finalized the Term Project by applying user testing and playing for edge cases. Arranged the hierarchy tab in such a way that all of the game objects are logically classified under related categories to enable observers to understand the relations between them. Changed the game ending screens for both winning and losing cases to make the game more realistic. In order to increase the user experience, added and modified some parts related to game objects concerning portals, collidable objects, and doors. This commit, hopefully, concludes all the required and extra parts for the term project.